pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["v-for-vinney"]
repository = "https://github.com/V-for-Vinney/python-project-lvl2"

packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
